convert this project to a typescript project
